+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
 ## UNRELEASED
+- Looking at a book placed in a Chiseled Bookshelf now shows what book it is.
+
+
 - Font color of the Book Signing Screen 'Enter Book Title' and 'by <author>' text is now configurable.
 - Font color of the book title text box now uses 'MainFontColor' config setting.